Separated file for services list tests created.
